Unstyled restuarant page with dummy content
This draft of the project has the required
functionality but it is unstyled and uses dummy
content.

Within /src there are three modules that
encapsulates functions that are used to generate
elements for and display a respective tab. The
index.js file consumes said modules and controls
the user interface.